chore: move related files to modules/user dir
