+++ frontend/src/modules/user/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Login.user'
+export * from './Register.user'
